+++ crates/grpc-server/grpc-server/src/server/disputes.rs
@@ -117,6 +117,7 @@ impl DisputeService for Disputes {
                         proxy_name,
                         tenant_id,
                         merchant_id,
+                        org_id,
                         connector_latency,
                         environment,
                         ..
@@ -184,6 +185,7 @@ impl DisputeService for Disputes {
                         proxy_name: proxy_name.as_deref(),
                         tenant_id: &tenant_id,
                         merchant_id: merchant_id.as_str(),
+                        org_id: org_id.as_str(),
                         return_raw_connector_data: config.common.return_raw_connector_data,
                         connector_latency,
                         log_fields_enabled: config.log_fields.enabled,
@@ -361,6 +363,7 @@ impl DisputeService for Disputes {
                         proxy_name,
                         tenant_id,
                         merchant_id,
+                        org_id,
                         connector_latency,
                         environment,
                         ..
@@ -429,6 +432,7 @@ impl DisputeService for Disputes {
                         proxy_name: proxy_name.as_deref(),
                         tenant_id: &tenant_id,
                         merchant_id: merchant_id.as_str(),
+                        org_id: org_id.as_str(),
                         return_raw_connector_data: config.common.return_raw_connector_data,
                         connector_latency,
                         log_fields_enabled: config.log_fields.enabled,