+++ crates/integrations/connector-integration/src/default_implementations.rs
@@ -292,6 +292,8 @@ default_impl_verify_webhook_source_v2!(
         Tesouro,
         Boost,
         Citigate,
+        Moneris,
+        Ilixium,
     ],
 );
 // PayPal has its own implementation in paypal.rs
@@ -370,6 +372,7 @@ macro_rules! default_impl_recharge_v2 {
 default_impl_recharge_v2!(
     Citigate,
     Boost,
+    Ilixium,
     Tesouro,
     AbsaSanlam,
     Aci,
@@ -420,6 +423,7 @@ default_impl_recharge_v2!(
     Loonio,
     Mifinity,
     Mollie,
+    Moneris,
     Multisafepay,
     Netcetera,
     Nexinets,
@@ -596,6 +600,7 @@ macro_rules! default_impl_get_payment_method_v2 {
 default_impl_create_payment_method_v2!(
     Citigate,
     Boost,
+    Ilixium,
     Tesouro,
     AbsaSanlam,
     Aci,
@@ -646,6 +651,7 @@ default_impl_create_payment_method_v2!(
     Loonio,
     Mifinity,
     Mollie,
+    Moneris,
     Multisafepay,
     Netcetera,
     Nexinets,
@@ -700,6 +706,7 @@ default_impl_create_payment_method_v2!(
 default_impl_get_payment_method_v2!(
     Citigate,
     Boost,
+    Ilixium,
     Tesouro,
     AbsaSanlam,
     Aci,
@@ -750,6 +757,7 @@ default_impl_get_payment_method_v2!(
     Loonio,
     Mifinity,
     Mollie,
+    Moneris,
     Multisafepay,
     Netcetera,
     Nexinets,
@@ -805,6 +813,7 @@ default_impl_payment_method_eligibility_v2!(
     not_supported: [
         Citigate,
         Boost,
+        Ilixium,
         Tesouro,
         Adyen,
         Authorizedotnet,
@@ -860,6 +869,7 @@ default_impl_payment_method_eligibility_v2!(
         Loonio,
         Mifinity,
         Mollie,
+        Moneris,
         Multisafepay,
         Nexinets,
         Nexixpay,
@@ -970,6 +980,7 @@ macro_rules! default_impl_refresh_payment_method_v2 {
 default_impl_refresh_payment_method_v2!(
     Citigate,
     Boost,
+    Ilixium,
     AbsaSanlam,
     Aci,
     Adyen,
@@ -1021,6 +1032,7 @@ default_impl_refresh_payment_method_v2!(
     Loonio,
     Mifinity,
     Mollie,
+    Moneris,
     Multisafepay,
     Netcetera,
     Nexinets,